+++ pluginv1/service.pb.go
@@ -1233,12 +1233,13 @@ const file_liapoldus_plugin_v1_service_proto_rawDesc = "" +
 	"\x14WebSocketMessageKind\x12&\n" +
 	"\"WEBSOCKET_MESSAGE_KIND_UNSPECIFIED\x10\x00\x12\x1f\n" +
 	"\x1bWEBSOCKET_MESSAGE_KIND_TEXT\x10\x01\x12!\n" +
-	"\x1dWEBSOCKET_MESSAGE_KIND_BINARY\x10\x022\x97\x04\n" +
+	"\x1dWEBSOCKET_MESSAGE_KIND_BINARY\x10\x022\xff\x04\n" +
 	"\rPluginService\x12O\n" +
 	"\bManifest\x12$.liapoldus.plugin.v1.ManifestRequest\x1a\x1d.liapoldus.plugin.v1.Manifest\x12[\n" +
 	"\fConfigSchema\x12(.liapoldus.plugin.v1.ConfigSchemaRequest\x1a!.liapoldus.plugin.v1.ConfigSchema\x12^\n" +
 	"\vConfigApply\x12'.liapoldus.plugin.v1.ConfigApplyRequest\x1a&.liapoldus.plugin.v1.ConfigApplyResult\x12U\n" +
-	"\bShutdown\x12$.liapoldus.plugin.v1.ShutdownRequest\x1a#.liapoldus.plugin.v1.ShutdownResult\x12K\n" +
+	"\bShutdown\x12$.liapoldus.plugin.v1.ShutdownRequest\x1a#.liapoldus.plugin.v1.ShutdownResult\x12f\n" +
+	"\rDispatchApply\x12).liapoldus.plugin.v1.DispatchApplyRequest\x1a*.liapoldus.plugin.v1.DispatchApplyResponse\x12K\n" +
 	"\x04Call\x12 .liapoldus.plugin.v1.CallRequest\x1a!.liapoldus.plugin.v1.CallResponse\x12T\n" +
 	"\x06Stream\x12\".liapoldus.plugin.v1.StreamMessage\x1a\".liapoldus.plugin.v1.StreamMessage(\x010\x01B7Z5github.com/Liapoldus/pluginprotocol/pluginv1;pluginv1b\x06proto3"
 
@@ -1281,10 +1282,12 @@ var file_liapoldus_plugin_v1_service_proto_goTypes = []any{
 	(*ConfigSchemaRequest)(nil),      // 21: liapoldus.plugin.v1.ConfigSchemaRequest
 	(*ConfigApplyRequest)(nil),       // 22: liapoldus.plugin.v1.ConfigApplyRequest
 	(*ShutdownRequest)(nil),          // 23: liapoldus.plugin.v1.ShutdownRequest
-	(*Manifest)(nil),                 // 24: liapoldus.plugin.v1.Manifest
-	(*ConfigSchema)(nil),             // 25: liapoldus.plugin.v1.ConfigSchema
-	(*ConfigApplyResult)(nil),        // 26: liapoldus.plugin.v1.ConfigApplyResult
-	(*ShutdownResult)(nil),           // 27: liapoldus.plugin.v1.ShutdownResult
+	(*DispatchApplyRequest)(nil),     // 24: liapoldus.plugin.v1.DispatchApplyRequest
+	(*Manifest)(nil),                 // 25: liapoldus.plugin.v1.Manifest
+	(*ConfigSchema)(nil),             // 26: liapoldus.plugin.v1.ConfigSchema
+	(*ConfigApplyResult)(nil),        // 27: liapoldus.plugin.v1.ConfigApplyResult
+	(*ShutdownResult)(nil),           // 28: liapoldus.plugin.v1.ShutdownResult
+	(*DispatchApplyResponse)(nil),    // 29: liapoldus.plugin.v1.DispatchApplyResponse
 }
 var file_liapoldus_plugin_v1_service_proto_depIdxs = []int32{
 	18, // 0: liapoldus.plugin.v1.CallRequest.grants:type_name -> liapoldus.plugin.v1.ActiveGrant
@@ -1309,16 +1312,18 @@ var file_liapoldus_plugin_v1_service_proto_depIdxs = []int32{
 	21, // 19: liapoldus.plugin.v1.PluginService.ConfigSchema:input_type -> liapoldus.plugin.v1.ConfigSchemaRequest
 	22, // 20: liapoldus.plugin.v1.PluginService.ConfigApply:input_type -> liapoldus.plugin.v1.ConfigApplyRequest
 	23, // 21: liapoldus.plugin.v1.PluginService.Shutdown:input_type -> liapoldus.plugin.v1.ShutdownRequest
-	4,  // 22: liapoldus.plugin.v1.PluginService.Call:input_type -> liapoldus.plugin.v1.CallRequest
-	16, // 23: liapoldus.plugin.v1.PluginService.Stream:input_type -> liapoldus.plugin.v1.StreamMessage
-	24, // 24: liapoldus.plugin.v1.PluginService.Manifest:output_type -> liapoldus.plugin.v1.Manifest
-	25, // 25: liapoldus.plugin.v1.PluginService.ConfigSchema:output_type -> liapoldus.plugin.v1.ConfigSchema
-	26, // 26: liapoldus.plugin.v1.PluginService.ConfigApply:output_type -> liapoldus.plugin.v1.ConfigApplyResult
-	27, // 27: liapoldus.plugin.v1.PluginService.Shutdown:output_type -> liapoldus.plugin.v1.ShutdownResult
-	5,  // 28: liapoldus.plugin.v1.PluginService.Call:output_type -> liapoldus.plugin.v1.CallResponse
-	16, // 29: liapoldus.plugin.v1.PluginService.Stream:output_type -> liapoldus.plugin.v1.StreamMessage
-	24, // [24:30] is the sub-list for method output_type
-	18, // [18:24] is the sub-list for method input_type
+	24, // 22: liapoldus.plugin.v1.PluginService.DispatchApply:input_type -> liapoldus.plugin.v1.DispatchApplyRequest
+	4,  // 23: liapoldus.plugin.v1.PluginService.Call:input_type -> liapoldus.plugin.v1.CallRequest
+	16, // 24: liapoldus.plugin.v1.PluginService.Stream:input_type -> liapoldus.plugin.v1.StreamMessage
+	25, // 25: liapoldus.plugin.v1.PluginService.Manifest:output_type -> liapoldus.plugin.v1.Manifest
+	26, // 26: liapoldus.plugin.v1.PluginService.ConfigSchema:output_type -> liapoldus.plugin.v1.ConfigSchema
+	27, // 27: liapoldus.plugin.v1.PluginService.ConfigApply:output_type -> liapoldus.plugin.v1.ConfigApplyResult
+	28, // 28: liapoldus.plugin.v1.PluginService.Shutdown:output_type -> liapoldus.plugin.v1.ShutdownResult
+	29, // 29: liapoldus.plugin.v1.PluginService.DispatchApply:output_type -> liapoldus.plugin.v1.DispatchApplyResponse
+	5,  // 30: liapoldus.plugin.v1.PluginService.Call:output_type -> liapoldus.plugin.v1.CallResponse
+	16, // 31: liapoldus.plugin.v1.PluginService.Stream:output_type -> liapoldus.plugin.v1.StreamMessage
+	25, // [25:32] is the sub-list for method output_type
+	18, // [18:25] is the sub-list for method input_type
 	18, // [18:18] is the sub-list for extension type_name
 	18, // [18:18] is the sub-list for extension extendee
 	0,  // [0:18] is the sub-list for field type_name